# system/src/main/resources/META-INF/openapi.yaml
openapi: 3.0.0
info:
  description: This is the REST contract for the Order System
  version: 1.0.0
  title: (20F) Group Project - Order System
  contact:
    name: (20F) CST 8277
    email: normanm@algonquincollege.com
    url: https://brightspace.algonquincollege.com/d2l/home/279461
servers:
  - url: http://localhost:{port}/{contextRoot}/{basePath}
    variables:
      port:
        default: "8080"
      contextRoot:
        default: rest-orderSystem
      basePath:
        default: api/v1
security:
  - basicAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
    idPathParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        description: a resource's id
        example: 1
...
  responses:
    ForbiddenError:
      description: Access to the specified resource has been forbidden
...
    UnauthorizedUserError:
      description: request requires HTTP authentication for USER_ROLE
    NotFound:
      description: The specified resource was not found
  schemas:
    PojoBase:
      type: object
      required:
        - id
        - version
      properties:
        id:
          description: id of entity
          type: integer
        version:
          description: version (OptLock) of entity
          type: integer
        createdDate:
          description: creation timestamp of entity
          type: string
          format: date-time
        updatedDate:
          description: last-updatd timestamp of entity
          type: string
          format: date-time
...
    Product:
      description: (20F) Group Project - Product entity
      type: object
      allOf:
        - $ref: "#/components/schemas/PojoBase"
      properties:
        description:
          type: string
          description: Description of the product
          example: Shirt
        serialNo:
          type: string
          description: serialNo of the product
          example: Abc123
    ProductsList:
      description: list of products
      type: array
      items:
        $ref: "#/components/schemas/Product"
    Store:
      description: (20F) Group Project - Store entity
      type: object
      allOf:
        - $ref: "#/components/schemas/PojoBase"
      properties:
        storeName:
          type: string
          description: Store Name
          example: Hudson Bay
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
    StoresList:
      description: list of stores
      type: array
      items:
        $ref: "#/components/schemas/Store"
paths:
...
  /product/{id}:
    get:
      summary: Retrieves the specified Product (by id)
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      responses:
        200:
          description: Ok (Successful operation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        401:
          $ref: "#/components/responses/UnauthorizedUserError"
        403:
          $ref: "#/components/responses/ForbiddenError"
        404:
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update the specified Product (by id)
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        200:
          description: Ok (Successful operation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsList"
...
  /store/{id}:
    get:
      summary: Retrieves the specified Store (by id)
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      responses:
        200:
          description: Ok (Successful operation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        401:
          $ref: "#/components/responses/UnauthorizedUserError"
        403:
          $ref: "#/components/responses/ForbiddenError"
        404:
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Removes the specified Store (by id)
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      responses:
        204:
          description: Ok (successful operation, no content)
        401:
          $ref: "#/components/responses/UnauthorizedUserError"
        403:
          $ref: "#/components/responses/ForbiddenError"
        404:
          $ref: "#/components/responses/NotFound"
    put:
      summary: Updates the specified Store (by id)
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      requestBody:
        description: update store to Db
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Store"
      responses:
        200:
          description: Ok (Successful operation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoresList"